moved C++ functions around to simplify organisation
