+++ setup.py
@@ -22,15 +22,16 @@ linear_algebra_module = Extension('pyml.maths.Clinear_algebra',
                                   include_dirs=['pyml/maths/include'],
                                   language='c++')
 
-gradient_descent_module = Extension('pyml.maths.gradient_descent',
-                                    sources=['pyml/maths/src/gradientdescentmodule.cpp',
-                                             'pyml/maths/src/linearalgebramodule.cpp',
-                                             'pyml/maths/src/pythonconverters.cpp',
-                                             'pyml/maths/src/flatArrays.cpp'],
-                                    extra_compile_args=['-std=c++11'],
-                                    # extra_link_args=['-lgomp'],
-                                    include_dirs=['pyml/maths/include'],
-                                    language='c++')
+optimisers_module = Extension('pyml.maths.optimisers',
+                              sources=['pyml/maths/src/optimisers.cpp',
+                                       'pyml/maths/src/gradientDescent.cpp',
+                                       'pyml/maths/src/linearalgebramodule.cpp',
+                                       'pyml/maths/src/pythonconverters.cpp',
+                                       'pyml/maths/src/flatArrays.cpp'],
+                              extra_compile_args=['-std=c++11'],
+                              # extra_link_args=['-lgomp'],
+                              include_dirs=['pyml/maths/include'],
+                              language='c++')
 
 distances = Extension('pyml.metrics.CMetrics',
                       sources=['pyml/metrics/src/metricspythonextension.cpp',
@@ -80,5 +81,5 @@ setup(
     author_email=about['__author_email__'],
     description='Machine learning with Python and C/C++',
     test_suite="tests",
-    ext_modules=[linear_algebra_module, gradient_descent_module, distances, maths],
+    ext_modules=[linear_algebra_module, optimisers_module, distances, maths],
 )
